Move the S3 fixture into the new storage lib
