+++ sbt_common/storage/src/test/scala/uk/ac/wellcome/storage/s3/S3ObjectStoreTest.scala
@@ -7,10 +7,10 @@ import org.scalatest.concurrent.ScalaFutures
 import org.scalatest.{FunSpec, Matchers}
 import uk.ac.wellcome.models.Sourced
 import uk.ac.wellcome.models.aws.S3Config
+import uk.ac.wellcome.storage.test.fixtures.S3
 import uk.ac.wellcome.test.utils.{ExtendedPatience, JsonTestUtil}
 import uk.ac.wellcome.utils.JsonUtil
 import uk.ac.wellcome.utils.JsonUtil._
-import uk.ac.wellcome.test.fixtures.S3
 
 import scala.concurrent.ExecutionContext.Implicits.global
