+++ sierra_adapter/sierra_reader/src/test/scala/uk/ac/wellcome/platform/sierra_reader/ServerTest.scala
@@ -4,7 +4,7 @@ import com.twitter.finagle.http.Status._
 import com.twitter.inject.server.FeatureTest
 import org.scalatest.FunSpec
 import uk.ac.wellcome.messaging.test.fixtures.SQS
-import uk.ac.wellcome.test.fixtures.S3
+import uk.ac.wellcome.storage.test.fixtures.S3
 
 class ServerTest extends FunSpec with fixtures.Server with S3 with SQS {
